pyproject.toml
[project]
name = "gmail-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.3",
    "mcp[cli]>=1.25.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
